Add: 비밀번호, 비밀번호 확인, 이메일 검증 구현
Server & Client

+++ core/verify_data.php
@@ -70,6 +70,35 @@ function verify_data($string, $mode) {
                     $data->outMsg = "<span class='text-danger'>이미 존재하는 닉네임입니다.</span>";
                 }
                 break;
+            case "password":
+                $data->outMsg = "<span class='text-success'>사용 가능한 비밀번호입니다.</span>";
+
+                if (strlen($string) < 8) {
+                    $data->verified = false;
+                    $data->outMsg = "<span class='text-danger'>비밀번호의 길이는 최소 8자이어야 합니다.</span>";
+                }
+
+                if (strlen($string) > 30) {
+                    $data->verified = false;
+                    $data->outMsg = "<span class='text-danger'>비밀번호의 길이는 최대 30자까지만 허용됩니다.</span>";
+                }
+                break;
+            case "password_confirm":
+                $data->outMsg = "<span class='text-success'>비밀번호가 일치합니다.</span>";
+
+                if (strcmp($string, $_POST["password"]) !== 0) { //not equal
+                    $data->verified = false;
+                    $data->outMsg = "<span class='text-danger'>비밀번호가 일치하지 않습니다.</span>";
+                }
+                break;
+            case "email":
+                $data->outMsg = "<span class='text-success'>사용 가능한 이메일입니다.</span>";
+
+                if (preg_match("/^[_\.0-9a-zA-Z-]+@([0-9a-zA-Z][0-9a-zA-Z-]+\.)+[a-zA-Z]{2,6}$/i", $string) == false) {
+                    $data->verified = false;
+                    $data->outMsg = "<span class='text-danger'>올바른 이메일 형식이 아닙니다.</span>";
+                }
+                break;
             default:
                 return false;
         }
@@ -84,6 +113,9 @@ function verify_data($string, $mode) {
 
 $data->user_id = verify_data($_POST["user_id"], "user_id");
 $data->nickname = verify_data($_POST["nickname"], "nickname");
+$data->pass = verify_data($_POST["password"], "password");
+$data->pass_confirm = verify_data($_POST["password_confirm"], "password_confirm");
+$data->email = verify_data($_POST["email"], "email");
 
 if ($_GET["mode"] === "ajax") {
     header("Content-Type:application/json");
@@ -92,7 +124,11 @@ if ($_GET["mode"] === "ajax") {
     unset($data);
 }
 else if ($_GET["mode"] === "process") {
-    if (!($data->user_id->verified && $data->nickname->verified)) {
+    if (!($data->user_id->verified &&
+        $data->nickname->verified &&
+        $data->pass->verified &&
+        $data->pass_confirm->verified &&
+        $data->email->verified)) {
         echo "<script>alert('비정상적인 접근이 감지되었습니다.');</script>";
         header('HTTP/1.1 301 Moved Permanently');
         header('Location: https://azure.mandora.xyz/dora-web/');
